vec4.h: initial commit
Initial commit of vec4.h module for math.h. With that addition, mat4.h module was enchanced with vec4.h module for some operations and sneak tricks. Lastly, common.h's sin(x) function was fixed (usage of trunc instead of fact, which in fact is a new addition for common.h module).