mat4.h: structure variable change
mat4 elements were changed from mat4::m0 - mat4::m15 to mat4::m00 - mat4::m33. This will allow more readability of the elements of the matrix and where each element lands as a representation of rows and columns. Implementation for the future mat2 and mat3 will follow the same convention. Also, 'float ptr[4][4]' was removed and will probably be replaced with a macro to the pointer to the first element of the matrix (helpful for working with OpenGL and, for example, passing data types into shaders).