common.h: pow(), exp() and log().
pow() function was rewritten to allow the usage of floating-point exponents for the operation. For that, exp(), log() and logn() functions were created. Following the addition, almost all easing functions work correctly. Lastly, styling of common.h functions was slightly changed and now all the functions should be prefixed with '_' and defined using macro (which will be used to select if function should be bound or not).