struct.h: initial commit.
Initial commit of new module: struct.h. It's specialized to define mathematical structures for the whole library. It was worth to modularize structures for the case of if user would like to use structures such as vectors with additional option of defining their own behaviour without clustering the codebase with current implementations (which due to the fact the whole library is static-inline'ed the final output can be quite bulky). What's more, mat4 now has it's C++ operators moved to the end of the file, the same as the other modules.