math.h: MATHAPI keyword.
From now on API can be (and this time I'll preserve this) an option to be either 'extern' (with MATHAPI_EXTERN) or 'static inline' (with MATHAPI_STATIC). With this change, common.h has been transformed to include both declaration and implementation secion; implementation can be either included with a new keyword: MATHAPI_IMPLEMENTATION or with a MATHAPI_STATIC keyword (which automatically define MATHAPI_IMPLEMENTATION keyword). Next I'll try to apply this format for all of the modules and then I'll try to make them more independent of each other.